Get length allocation working
Give up on `&mut slab::Index`-based logic for now, because it's too
hard to integrate `slab::Index: !Copy` and (de)-serialization logic
in the transfer protocol. The only place that might benefit from
mutable vs. atomic access is in `block::Set`, which we can special
case later with unsafe.