Keep count instead of full sparse field
The latter actually maintains more information than we need,
i.e. the position of the full blocks. For now, we'll get rid
of it to speed up the `is_full` query.

Forgot this was an option since keeping the count up-to-date for
the intrusive linked list was more complex/expensive (parity bit).
We do still need the partial sparse field to accelerate the peek
function, which does require the position of blocks.