Instrument transfer protocol
Not sure if we want to keep this in the same diagram as the allocate
and free flows, which make more sense. Here it's possible for contention
counters to grow to #threads * #attempts, if everyone contends on
everyone else's operation, whereas the allocate and free counters
are actually disjoint and sum to the parent counter.