Reuse log entries
Allocation entries are safe to overwrite after the corresponding
free is committed. Free entries are (conservatively?) safe to
overwrite after every process has acknowledged and unmapped the free.
Not sure if we can be even lazier and just asynchronously unlink
and unmap any existing mappings that are not recorded in the log?
Then free can be reused immediately after being committed + marking
the original allocation entry.

Because we use an allocating interval set in DRAM to keep track of
available offsets, this implicitly creates a `Drop` implementation
for `Allocator`. Rust uses __cxa_thread_atexit for its implementation
of native thread local destructors, which is *not* interposed by
the global allocator, and calls the libc API directly.

To work around this, we use `ManuallyDrop` and our existing pthread
key infrastructure so that there is no allocation or deallocation
happening in the TLS destructors.

Still need to implement LSN wraparound, retries on conflict, and
recovery logic.