Replace `AtomicRefCell` with specialized lock
The problem with the former is that it would only lock the slab
and identify conflict during a local allocate or free operation,
which is extremely fast.

Instead, implement our own lock that tracks the thread ID, and
lock slab ownership at every coarse-grained transition. Also add
extra debugging information about the state of the slab metadata.
The lock operations are a no-op without the `validate` feature flag.

This helped identify a bug where ownership was not updated correctly
for slabs that were detached (but not disowned), and eventually
entirely freed by remote threads, which assumed the slab had no owner.

Another bug had to do with our remote free CAS, which must use
the metadata version *before* making our bit globally visible,
to guarantee that it could not have changed.