Implement parametrized slab region
Interleave the owned and shared metadata to make heap extension
simpler for now. Not sure what the best way to make this configurable
is, so we can measure the performance impact--changing it back would
require mapping two different regions for the two kinds of descriptors,
i.e. either Map((Owned, Shared)) or (Map(Owned), Map(Shared)).