Move class from remote to local
This allows us to use a 16-bit version again, but has the downside
that we can no longer check the size class of an arbitrary allocation.
This will break APIs like realloc and malloc_usable_size, since
the local slab descriptors can be cached and not necessarily up-to-date
in memory.