Fix subtle reference counting race condition
When we successfully `try_swap` in the fast path, the reference
count of the value is 1 from the root reference. If another thread
interleaves after we publish the value, but before we increment
its reference count, that thread can unlink, setting the reference
count to 0, and free before we finally link the value.

The fix is to increment the reference count before we publish the value.